requirements.txt
Flask
google-generativeai
python-dotenv
transformers
torch
sentencepiece